Remove ignore of runtime.goexit for Ltree
When the Ltree flag is used in testing, the runtime.goexit() function is in the
call stack and this increases the indent level of all output. This block of
code pushed back the indent level so it looked better, but was incorrect. Let's
leave it like this for a while and see how it fits.